feature/conn25: move addrAssignments to their own file
Updates tailscale/corp#39975
